190918 django_CRUD_adv

+++ 03_django/02_django_crud/articles/views.py
@@ -12,12 +12,9 @@ def index(request):
     return render(request, 'articles/index.html', context)
 
 
-def new(request):
-    return render(request, 'articles/new.html')
-
-
 def create(request):
-    try:
+    # CREATE
+    if request.method == 'POST':
         title = request.POST.get('title')
         content = request.POST.get('content')
 
@@ -29,17 +26,17 @@ def create(request):
 
         # 2
         article = Article(title=title, content=content)
-        article.full_clean()
-    except ValidationError:
-        raise ValidationError('Error')
-    else:
+    
         article.save()
 
         # 3
         # Article.objects.create(title=title, content=content)
 
-        return redirect(f'/articles/{article.pk}/') # 메인 페이지
+        return redirect(article) # 메인 페이지
         # return redirect('/articles/', article.pk)
+    # NEW
+    else:
+        return render(request, 'articles/create.html')
 
 
 def detail(request, pk):
@@ -50,19 +47,20 @@ def detail(request, pk):
 
 def delete(request, pk):
     article = Article.objects.get(pk=pk)
-    article.delete()
-    return redirect('/articles/')
-
-
-def edit(request, pk):
-    article = Article.objects.get(pk=pk)
-    context = {'article': article,}
-    return render(request, 'articles/edit.html', context)
+    if request.method == 'POST':
+        article.delete()
+        return redirect('articles:index')
+    else:
+        return redirect(article)
 
 
 def update(request, pk):
     article = Article.objects.get(pk=pk)
-    article.title = request.POST.get('title')
-    article.content = request.POST.get('content')
-    article.save()
-    return redirect(f'/articles/{article.pk}/')
+    if request.method == 'POST':        
+        article.title = request.POST.get('title')
+        article.content = request.POST.get('content')
+        article.save()
+        return redirect(article)
+    else:
+        context = {'article': article,}
+        return render(request, 'articles/update.html', context)
